scripts/requirements.txt
requests
urllib3
pytz
google-auth
pandas
openpyxl
google-cloud-bigquery